chore: separating new architecture from current again and adapting tests.
