+++ app/src/main/java/com/android/shelfLife/model/newhousehold/HouseHoldRepositoryFirestore.kt
@@ -114,7 +114,6 @@ class HouseholdRepositoryFirestore(
       updateSelectedHousehold()
     } catch (e: Exception) {
       Log.e("HouseholdRepository", "Error initializing households", e)
-      _households.value = emptyList()
     }
   }
