+++ app/src/main/java/com/android/shelfLife/ui/newoverview/NewFirstTimeWelcome.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.android.shelfLife.model.overview.OverviewScreenViewModel
+import com.android.shelfLife.model.newOverview.OverviewScreenViewModel
 import com.android.shelfLife.ui.navigation.NavigationActions
 import com.android.shelfLife.ui.navigation.Route
 import com.android.shelfLife.ui.navigation.Screen
