+++ app/src/main/java/com/android/shelfLife/ui/newoverview/NewHouseHoldCreationScreen.kt
@@ -33,14 +33,14 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.android.shelfLife.R
-import com.android.shelfLife.model.creationScreen.CreationScreenViewModel
+import com.android.shelfLife.model.newCreationScreen.CreationScreenViewModel
 import com.android.shelfLife.model.newInvitations.InvitationRepository
 import com.android.shelfLife.model.newhousehold.HouseHoldRepository
 import com.android.shelfLife.model.user.UserRepository
 import com.android.shelfLife.ui.navigation.NavigationActions
 import com.android.shelfLife.ui.navigation.Screen
+import com.android.shelfLife.ui.newutils.DeletionConfirmationPopUp
 import com.android.shelfLife.ui.utils.CustomButtons
-import com.android.shelfLife.ui.utils.DeletionConfirmationPopUp
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.launch
