+++ app/src/test/java/com/android/shelflife/model/newInvitations/InvitationRepositoryFirestoreTest.kt
@@ -90,7 +90,7 @@ class InvitationRepositoryFirestoreTest {
     `when`(mockDocument.id).thenReturn(invitationId)
     `when`(mockDocument.set(any<Map<String, Any>>())).thenReturn(mockTask)
 
-    invitationRepository.sendInvitation(household, invitedUser)
+    invitationRepository.sendInvitation(household, invitedUser.uid)
 
     val invitationDataCaptor = ArgumentCaptor.forClass(Map::class.java)
     verify(mockDocument).set(invitationDataCaptor.capture())
